requirements.txt
Django==2.0
django-filter==1.1.0
django-filters==0.2.1
djangorestframework==3.7.4
djangorestframework-filters==0.10.2
pytz==2017.3
